requirements.txt
# Travel Agent - Python dependencies
# Install: pip install -r requirements.txt

# Environment
python-dotenv>=1.0.0

# HTTP
requests>=2.28.0

# LangChain & LangGraph
langchain-core>=0.3.0
langchain-groq>=0.2.0
langgraph>=0.2.0

# Web app
streamlit>=1.28.0
